[IMP] iap, sms: improve sms registration
Refactor the iap_account and iap_service models, which allows for a
better registration flow for sms accounts. Where the user can verify
their phone number, set a sender name and warning emails on their own db
without needing to go to the iap website. Additionaly the
assertQueryCount had to be increased for some tests, as the iap.account
is set outside of _contact_iap, which is mocked in those tests.

task-3741280

closes odoo/odoo#175562

Related: odoo/enterprise#67883
Related: odoo/upgrade#6346
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>